⏪️ Put the USB connector back down (recessed)
Because most cables used with the keyboard will be straights (and not right angled), I prefer to just recess the connector and keep 2 mm margin on each side for the male connector to fit in. That makes the PCB more rigid overall (compared to having that 2 mm-wide finger for the through-hole pads of the connector), but of course prevents the use of most right-angled connectors.