🚨 Fix all DRC issues!
For real! Still have to check clearance to board outline later on, mostly for the jack connector’s pads.
Other half can probably be started now! Or maybe I try to make this one reversible, that would be cool, but required two-sides assembly because of the LEDs.
This evening was pretty productive, good night!